Added this for notifications, read description.
afk=True was only in utils.py/presence command. This made it so that once the bot restarts (i.e. every 24h), or if the user does not do the command, they get no notifications. This should fix it.

Submitting as a pull request so verixx can see it first.